refactor: 重构ContestList，组件解耦
